Recipe 등록수정삭제 기능

+++ Recipe/src/main/java/com/spring/biz/view/recipe/RecipeController.java
@@ -7,6 +7,8 @@ import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import com.spring.biz.blog.BlogVO;
 import com.spring.biz.member.MemberVO;
@@ -23,26 +25,53 @@ public class RecipeController {
 	}
 	
 	@RequestMapping(value = "/insertRecipe.do")
-	public String insertRecipe(RecipeVO rvo) {
+	public String insertRecipe(RecipeVO rvo, HttpSession sess) {
 		System.out.println("===Controller의 insertRecipe() 실행===");
 		System.out.println("rvo : " + rvo);
+		MemberVO mvo = (MemberVO) sess.getAttribute("member");
+		rvo.setWritten(mvo.getNickname());
+		rvo.setId(mvo.getId());
 		int result = recipeService.insertRecipe(rvo);
-		return "상세레시피화면.jsp";
+		return "recipeSingle";
+	}
+	
+	@RequestMapping(value = "/updateRjsp.do", method=RequestMethod.GET)
+	public String updateRjsp(@RequestParam int recipeno, Model model, HttpSession sess) {
+		System.out.println("===Controller의 updateRjsp() 실행===");
+		MemberVO mvo = (MemberVO) sess.getAttribute("member");
+		String logid = mvo.getId();
+		String recipeid = recipeService.idChk(recipeno);
+		if(logid.equals(recipeid)) {
+			RecipeVO rvo = recipeService.getRecipe(recipeno);
+			model.addAttribute("recipe", rvo);
+			return "recipeUpdate";
+		}else {
+			return "redirect:recipeSingle";
+		}
 	}
 	
 	@RequestMapping(value = "/updateRecipe.do")
-	public String updateRecipe(RecipeVO rvo) {
+	public String updateRecipe(RecipeVO rvo, Model model) {
 		System.out.println("===Controller의 updateRecipe() 실행===");
 		System.out.println("rvo : " + rvo);
 		int result = recipeService.updateRecipe(rvo);
-		return "상세레시피화면.jsp";
+		RecipeVO recipe = recipeService.getRecipe(rvo.getRecipeno());
+		model.addAttribute("recipe", recipe);
+		return "recipeSingle";
 	}
 	
-	@RequestMapping(value = "/deleteRecipe.do")
-	public String deleteRecipe(int recipeno) {
+	@RequestMapping(value = "/deleteRecipe.do", method=RequestMethod.GET)
+	public String deleteRecipe(@RequestParam int recipeno, HttpSession sess) {
 		System.out.println("===Controller의 deleteRecipe() 실행===");
-		recipeService.deleteRecipe(recipeno);
-		return "전체레시피화면.jsp";
+		MemberVO mvo = (MemberVO) sess.getAttribute("member");
+		String logid = mvo.getId();
+		String recipeid = recipeService.idChk(recipeno);
+		if(logid.equals(recipeid)) {
+			recipeService.deleteRecipe(recipeno);
+			return "Recipe";
+		}else {
+			return "redirect:recipeSingle";
+		}
 	}
 	
 	@RequestMapping(value = "/getRecipeList.do")
@@ -63,10 +92,11 @@ public class RecipeController {
 		return "myRecipe";
 	}
 	
-	@RequestMapping(value = "/getRecipe.do")
-	public String getRecipe(int recipeno) {
+	@RequestMapping(value = "/getRecipe.do", method=RequestMethod.GET)
+	public String getRecipe(@RequestParam int recipeno, Model model) {
 		System.out.println("===Controller의 getRecipe() 실행===");
 		RecipeVO rvo = recipeService.getRecipe(recipeno);
-		return "상세레시피화면.jsp";
+		model.addAttribute("recipe", rvo);
+		return "recipeSingle";
 	}
 }
